+++ src/main/java/kitchenpos/domain/menu/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.domain.menu.repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
