+++ src/main/java/kitchenpos/domain/menu/repository/MenuProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.domain.menu.repository;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
