+++ src/main/java/kitchenpos/domain/menu/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.domain.menu.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
