+++ src/main/java/kitchenpos/domain/order/repository/OrderTableRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.domain.order.repository;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
