+++ src/main/java/kitchenpos/domain/order/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.domain.order.repository;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
