+++ src/test/java/kitchenpos/application/fixture/MenuProductFixture.java
@@ -1,8 +1,8 @@
 package kitchenpos.application.fixture;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.Product;
 
 public abstract class MenuProductFixture {
