+++ src/test/java/kitchenpos/application/fixture/OrderFixture.java
@@ -1,9 +1,9 @@
 package kitchenpos.application.fixture;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.order.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
